requirements.txt
discord
Flask
gunicorn
line-bot-sdk
python-dotenv